requirements.txt
fastapi~=0.115.6
uvicorn~=0.32.1
httpx~=0.28.1
pydantic~=2.10.3
python-dotenv~=1.0.1
sqlalchemy~=2.0.36
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
starlette~=0.41.3
bcrypt
python-multipart
